Tweaked constraints to increase room variety
Made it so inner-walls can be backed by outer-walls. This doesn't
initially seem desirable, but a quirk of constraint propagation was
preventing inner-walls from connecting to diagonal walls.

The result was more longer/narrow rooms in large bubbles. The change
here makes it so bubbles have a larger variety of rooms at the cost of
some inner-wall-on-outer-wall artifacts.