Fixed incremental generation by very naive/rough collision detection
Needs line-circle collision detection, currently just testing every
pixel on the line against our bubble radius.

This could be faster with some math. And we may be able to take
advantage of the fact these are only aligned lines to simplify things.

Then again do we really care if we're still constrained by wfc? Maybe
move this into gen_bubbles so the time measurement is in the right
place.